Rename gridadapter.py to xgrid.py
