+++ tests/v4/test_xgrid.py
@@ -7,8 +7,8 @@ from numpy.testing import assert_allclose
 from parcels import xgcm
 from parcels._datasets.structured.generic import T, X, Y, Z, datasets
 from parcels.grid import Grid as OldGrid
-from parcels.gridadapter import XGrid
 from parcels.tools.converters import TimeConverter
+from parcels.xgrid import XGrid
 
 GridTestCase = namedtuple("GridTestCase", ["Grid", "attr", "expected"])
 
@@ -38,7 +38,7 @@ def assert_equal(actual, expected):
 
 
 @pytest.mark.parametrize("ds, attr, expected", test_cases)
-def test_grid_adapter_properties_ground_truth(ds, attr, expected):
+def test_xgrid_properties_ground_truth(ds, attr, expected):
     adapter = XGrid(xgcm.Grid(ds, periodic=False))
     actual = getattr(adapter, attr)
     assert_equal(actual, expected)
@@ -60,7 +60,7 @@ def test_grid_adapter_properties_ground_truth(ds, attr, expected):
     ],
 )
 @pytest.mark.parametrize("ds", datasets.values())
-def test_grid_adapter_against_old(ds, attr):
+def test_xgrid_against_old(ds, attr):
     adapter = XGrid(xgcm.Grid(ds, periodic=False))
 
     grid = OldGrid.create_grid(
